Nova/backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
python-dotenv>=1.0.0

# Note: Klix dependencies (google-genai, mem0, ollama, etc.) 
# should already be installed from the parent project
